backend/requirements.txt
absl-py==1.2.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
audioread==3.0.1
backcall==0.2.0
bokeh==3.1.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==5.2.1
dill==0.3.9
dm-sonnet==2.0.0
dm-tree==0.1.8
etils==1.3.0
executing==2.2.0
flatbuffers==1.12
fonttools==4.56.0
future==1.0.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.63.1
grpcio==1.70.0
h5py==3.11.0
idna==3.10
imageio==2.20.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
intervaltree==3.1.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
librosa==0.7.2
llvmlite==0.32.1
magenta==2.1.4
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.5.2
matplotlib-inline==0.1.7
mido==1.2.6
mir_eval==0.7
networkx==3.1
note-seq==0.0.3
numba==0.49.1
numpy==1.21.6
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
Pillow==9.2.0
pretty_midi==0.2.9
promise==2.3
prompt_toolkit==3.0.50
protobuf==3.19.6
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydub==0.25.1
Pygments==2.19.1
pygtrie==2.5.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-rtmidi==1.1.2
pytz==2025.1
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.7.3
six==1.16.0
sk-video==1.1.10
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.4.1
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-datasets==4.6.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.13.0
tensorflow-probability==0.17.0
termcolor==2.4.0
tf-slim==1.1.0
threadpoolctl==3.5.0
tifffile==2023.7.10
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
wrapt==1.17.2
xyzservices==2025.1.0
zipp==3.20.2
